Replace all occurrences of size_t with uint32_t

+++ src/include/stdlib.c
@@ -36,8 +36,8 @@ char *itoa(int value, char *str, int base) {
   return rc;
 }
 
-void *memcpy(char *dest, const char *src, size_t size) {
-  for (size_t i = 0; i < size; i++)
+void *memcpy(char *dest, const char *src, uint32_t size) {
+  for (uint32_t i = 0; i < size; i++)
     dest[i] = src[i];
   return dest;
 }
