+++ src/include/string.h
@@ -1,9 +1,10 @@
 #ifndef STRING_H
 #define STRING_H
 
-#include <stddef.h>
-size_t strlen(const char *str);
+#include "stdint.h"
+
+uint32_t strlen(const char *str);
 int strcmp(const char *str1, const char *str2);
-size_t trim(char *out, size_t len, const char *str);
+uint32_t trim(char *out, uint32_t len, const char *str);
 
 #endif
